requirements.txt
selenium
rich
PyCryptodomex
cx_Freeze
dload